8.20 / 21 - add items to your game
Added functionality for items within the game. Implemented an Item class that has an itemWeight and name. The Room class has been extended to include a room item field in each room.

In each room, when you use the look command, if there are any items or things to take available, the game will display information about them.

The item class should produce the string decription and the game class will call to print this info.